deleted duplicate folder and cleaned up structure
