+++ diamond_data_react folder/src/types/game.ts
@@ -1,0 +1,11 @@
+export type GameStats = {
+  atBats: number;
+  hits: number;
+  singles: number;
+  doubles: number;
+  triples: number;
+  homeRuns: number;
+  rbi: number;
+  walks: number;
+  strikeOuts: number;
+}
