+++ diamond_data_react folder/src/types/position.ts
@@ -1,0 +1,1 @@
+export type PlayerPosition = 'P' | 'C' | '1B' | '2B' | '3B' | 'SS' | 'OF'
