+++ diamond_data_react folder/src/utils/stats.ts
